+++ src/app/(route)/customer/_types/NoticeInfoItemType.ts
