+++ src/app/(route)/customer/_types/NoticeItemType.ts
