+++ src/app/(route)/customer/feedback/feedback-info/[feedbackId]/page.tsx
@@ -19,9 +19,9 @@ import CommentBar from "@/app/_components/_gnb/_components/CommentBar";
 import PostAction from "@/app/(route)/(community)/_components/PostAction";
 import FeedbackItem from "../../../_components/FeedbackItem";
 import FeedbackItemSkeleton from "../../../_components/FeedbackItemSkeleton";
-import { FeedbackContentType } from "@/app/_constants/customer/FeedbackItemType";
+import { FeedbackContentType } from "@/app/(route)/customer/_types/FeedbackItemType";
 import useGetNoticeDataList from "@/_hooks/fetcher/customer/useGetNoticeDataList";
-import { NoticeContentType } from "@/app/_constants/customer/NoticeItemType";
+import { NoticeContentType } from "@/app/(route)/customer/_types/NoticeItemType";
 import NoticeItem from "../../../_components/NoticeItem";
 import RecommendButton from "@/app/(route)/(community)/_components/RecommendButton";
 import usePostFeedbackRecommend from "@/_hooks/fetcher/customer/Recommend/usePostFeedbackRecommend";
